set delete loading indicator

+++ app/_lib/actions.js
@@ -3,6 +3,7 @@
 import { revalidatePath } from "next/cache";
 import { auth, signIn, signOut } from "./auth";
 import { supabase } from "./supabase";
+import { getBookings } from "./data-service";
 
 export async function updateGuest(formData) {
   const session = await auth();
@@ -16,7 +17,7 @@ export async function updateGuest(formData) {
 
   const updateData = { nationalID, nationality, countryFlag };
 
-  const { data, error } = await supabase
+  const { error } = await supabase
     .from("guests")
     .update(updateData)
     .eq("id", session.user.guestId);
@@ -30,12 +31,19 @@ export async function deleteReservation(bookingId) {
   const session = await auth();
   if (!session) throw new Error("You must be logged in");
 
+  const bookings = await getBookings(session.user.guestId);
+  const bookingIds = bookings.map((booking) => booking.id);
+  if (!bookingIds.includes(bookingId))
+    throw new Error("You are not allowed to delete this booking");
+
   const { error } = await supabase
     .from("bookings")
     .delete()
     .eq("id", bookingId);
 
   if (error) throw new Error("Booking could not be deleted");
+
+  revalidatePath("/accounts/reservations");
 }
 
 export async function signInAction() {
